+++ src/main/java/insane96mcp/iguanatweaksexpanded/data/generator/ITEDamageTypeTagsProvider.java
@@ -1,7 +1,7 @@
 package insane96mcp.iguanatweaksexpanded.data.generator;
 
 import insane96mcp.iguanatweaksexpanded.IguanaTweaksExpanded;
-import insane96mcp.iguanatweaksexpanded.module.experience.enchantments.enchantment.CurseOfMending;
+import insane96mcp.iguanatweaksexpanded.module.experience.enchantments.enchantment.CurseOfExperience;
 import insane96mcp.iguanatweaksexpanded.module.items.copper.CopperToolsExpansion;
 import insane96mcp.iguanatweaksreborn.module.combat.PiercingPickaxes;
 import net.minecraft.core.HolderLookup;
@@ -27,9 +27,9 @@ public class ITEDamageTypeTagsProvider extends DamageTypeTagsProvider {
     protected void addTags(HolderLookup.Provider pProvider) {
         tag(PiercingPickaxes.DOESNT_TRIGGER_PIERCING).add(CopperToolsExpansion.ELECTROCUTION_ATTACK);
 
-        tag(CopperToolsExpansion.DOESNT_TRIGGER_ELECTROCUTION).addOptionalTag(PiercingPickaxes.PIERCING_DAMAGE_TYPE.location()).add(CopperToolsExpansion.ELECTROCUTION_ATTACK).add(CurseOfMending.DAMAGE_TYPE);
+        tag(CopperToolsExpansion.DOESNT_TRIGGER_ELECTROCUTION).addOptionalTag(PiercingPickaxes.PIERCING_DAMAGE_TYPE.location()).add(CopperToolsExpansion.ELECTROCUTION_ATTACK).add(CurseOfExperience.DAMAGE_TYPE);
 
-        tag(DamageTypeTags.BYPASSES_ARMOR).add(CurseOfMending.DAMAGE_TYPE);
+        tag(DamageTypeTags.BYPASSES_ARMOR).add(CurseOfExperience.DAMAGE_TYPE);
     }
 
     public static TagKey<DamageType> create(String tagName) {