+++ changelog.md
@@ -1,11 +1,9 @@
 # Changelog
 
-* Curses on a tool increase the enchanting power, at the cost of the curse
-
 ## Upcoming
 * Curses!
   * Items with curses only can now be enchanted and get a 40% boost to max cost
-  * Added Curse of Mending. Consumes experience when used
+  * Added Curse of Experience. Consumes experience when used
 * Mining charge crafting is now cheaper
 * All items in creative tabs are now sorted!
 * Fixed beeg ore veins generating less at deepslate level